Moved kotlin functionality to separate plugin
