+++ plugins/de.cognicrypt.staticanalyzer.kotlin/META-INF/MANIFEST.MF
@@ -1,0 +1,18 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: CogniCrypt Static Analyzer for Kotlin
+Bundle-SymbolicName: de.cognicrypt.staticanalyzer.kotlin
+Bundle-Version: 1.0.0.qualifier
+Bundle-Activator: de.cognicrypt.staticanalyzer.kotlin.Activator
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Automatic-Module-Name: de.cognicrypt.staticanalyzer.kotlin
+Bundle-ActivationPolicy: lazy
+Bundle-ClassPath: .,
+ lib/
+Require-Bundle: org.eclipse.jdt.core,
+ org.eclipse.m2e.maven.runtime,
+ org.eclipse.core.runtime,
+ org.eclipse.core.resources,
+ org.eclipse.ui,
+ org.jetbrains.kotlin.core;bundle-version="0.8.19";resolution:=optional
+Export-Package: de.cognicrypt.staticanalyzer.kotlin.utilities
