+++ plugins/de.cognicrypt.staticanalyzer.kotlin/build.properties
@@ -1,0 +1,7 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               src/,\
+               lib/
+source.. = src/
