+++ plugins/de.cognicrypt.staticanalyzer/.classpath
@@ -3,6 +3,5 @@
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
 	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="lib" path="resources/org.jetbrains.kotlin.core.jar"/>
 	<classpathentry kind="output" path="target/classes"/>
 </classpath>
